Moved url to constants file
