--- src/main/webapp/creatives/maps/constants.js
@@ -1,11 +1,0 @@
-
-/** A list of server routes that keep track of each interaction when fetched. */
-export const Urls = {
-  BASE_URL: "https://step-capstone-team33-2020.appspot.com",
-  INITIALIZE: "initialize",
-  FIND_NEAREST_LOCATION: "find-location",
-  GRANT_LOCATION_DATA: "grant-location",
-  MAP_INTERACT: "map-interact",
-  SKIP_TO_CONTENT: "skip-to-content",
-  RETURN_TO_AD: "return-to-ad",
-}
